Enable UserResource in airlines app
Currently, the TCK verifies that the implementation creates paths and
operations for UserResource, even though it isn't configured as a rest
resource.

The spec doesn't demand that this should or shouldn't happen and it
makes sense for implementations to be allowed to check which resources
are actually included in the rest application and only generate schema
for those that are.