Allow annotation properties to wrap
The default MicroProfile formatting rules currently don't allow
annotation properties to wrap on a new line, which makes it hard to
maintain the TCK test applications because MP OpenAPI contains large
annotations with lots of attributes.

This change overrides the default formatter config XML file for the tck
project with one which has an additional property, allowing annotation
properties to wrap.