Make most archives not testable
A "testable" archive is one that arquillian adds extra framework code to
in order to run tests within the server runtime environment.

Most of the MP OpenAPI test classes don't include any tests which run on
the server side, they deploy an app and then query the /openapi endpoint
from the client side.

In addition, it causes problems when using the new arquillian REST 3.0
protocol. When using this protocol, an additional REST resource is
deployed within testable deployments which is scanned and added to the
openapi documentation which causes some of the tests to fail.